Fix frontend workflow
We have to break down the processing of agents in two phases.
first we map their attributes. This allows us to add stigmergies and
global processes to the symbols table without incurring in "undefined
reference" errors. Finally we can add agents' processes.

We also change mapVar so it operates on variables within the symbols
table, which have no external variable references.