Add BExpr.canonical, Bexpr.simplify
BExpr.simplify merges nested compound nodes with the same operator.
E.G. This BExpr:
(and (b1 (and b2 b3)))
becomes
(and b1 b2 b3)

BExpr.canonical uses the structural comparison of Expr type to rearrange
expressions within a BExpr (when it is safe to do so). This allows us to
recognize and remove multiple occurrence of the same sub-predicate within
a larger one (we do that within BExpr.simplify, via a List.distinctBy)