Move translateKit constructor to Common.fs
Most of the logic for translators is language-independent. You
only need to specify "elementary" translators (for (B)Exprs, Locations,
and LinkId) for a given language.
We cannot pass these translators as function arguments to translateKit,
as their types would get statically resolved. E.g. if we use C.translate
to build a (Expr<_, unit> -> string) translator, we cannot use it to
to build a (BExpr<_, LinkComponent> -> string) translator as well.
The workaround is to create an interface, as the members of its instances
will maintain genericity when passed as arguments.
(Source: https://stackoverflow.com/questions/42598677/what-is-the-best-way-to-pass-generic-function-that-resolves-to-multiple-types)