Refactor scheduler
Manage the scheduler within process step.
In fair systems, we only need to store the id of the last agent
that performed a transition. We do that with a "lastAgent"
process.

Add an existsEnabledAgent() function that checks if there
is at least 1 agent with no pending messages.

Add a "firstAgent" variable (for the active agent's id)