Fix LNT template
* Improve step process

* Deadlock detection

If:
a. There is at least 1 agent that has not terminated AND
b. No agent can perform a transition AND
c. There are no pending messages AND

Then we are in a "spurious" state (e.g. one that is not reachable
as per the semantics of LAbS). In C, these states are pruned away
by assume() statements; in LNT we have to deal with them.
Then, we emit a special "deadlock" message on the "monitor" gate, and
use the formula FAIRLY("MONITOR !TRUE" or "MONITOR !\"deadlock\") to
check liveness properties. (Safety properties are not affected by these
states).