Add block-local variables to LAbS syntax.
Action "foo := expr" evaluates expression expr and
stores the result in a variable "foo" that is only
available within the current block.
Thus, block
{ foo := 2*3; bar <- 4*foo }
will set "bar" to 12.
At the moment, accessing a variable before assigning
to it does not raise any warning/error; rather, the
variable is assumed to have nondeterministic value.
(This may change in the future).