Add "pick" primitive to the language
Within a block, action
x := pick E
(where E is a constant expression that evaluates to
a positive integer N) will initialize x as an array
of N id. The ids are guaranteed to be all different
from each other and from that of the agent performing
the action.

Caveat: N must be strictly less than (number of agents
in the system) - 1. (This is not checked yet)