Add "foreach" construct (assume/check)
This construct allows to predicate on every element of an array.
Syntax:
    foreach x in arr, bexpr(x)

Where bexpr(x) is a Boolean expression featuring x as a free variable.
This is translated into bexpr(arr[0]) and ... and bexpr(arr[n]); it
happens during tryAddProperty/tryAddAssume, so at the moment it is not
supported elsewhere (e.g., in processes).